[FIX]Fixed endpoints and bugs

+++ app/models/ballot.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column, ForeignKey, Index, Integer, String, Date,Numeric
+from sqlalchemy import Column, ForeignKey, Index, Integer, String, Date
+from sqlalchemy.types import BigInteger
 from sqlalchemy.orm import relationship
 from app.models.base import Base
 
@@ -8,7 +9,7 @@ class Ballot(Base):
     ballot_id = Column(Integer, primary_key=True, autoincrement=True)
     user_id = Column(Integer, ForeignKey('participants.user_id'), nullable=False)
     lottery_id = Column(Integer, ForeignKey('lotteries.lottery_id'), nullable=False)
-    ballot_number = Column(Integer, unique=True)
+    ballot_number = Column(BigInteger, unique=True)
     expiry_date = Column(Date)
 
     users = relationship("Participant", back_populates="ballots")
